Refactor project into multimodule project
Part of https://github.com/vaadin/flow-component-base/issues/117
