--- src/test/java/com/vaadin/flow/component/combobox/test/AutoFocusFilterIT.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2000-2018 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.combobox.test;
-
-import com.vaadin.flow.component.combobox.ComboBoxElementUpdated;
-import com.vaadin.flow.component.combobox.testbench.ComboBoxElement;
-import com.vaadin.flow.testutil.TestPath;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.By;
-
-@TestPath("auto-focus-filter")
-public class AutoFocusFilterIT extends AbstractComboBoxIT {
-
-    @Before
-    public void init() {
-        open();
-        waitUntil(driver -> findElements(By.tagName("vaadin-combo-box"))
-                .size() > 0);
-    }
-
-    @Test
-    public void filter_itemsShouldBeThere() {
-        ComboBoxElement comboBox = $(ComboBoxElementUpdated.class).first();
-
-        comboBox.sendKeys("2");
-
-        waitForItems(comboBox,
-                items -> items.size() == 2
-                        && "Option 2".equals(getItemLabel(items, 0))
-                        && "Another Option 2".equals(getItemLabel(items, 1)));
-    }
-
-}
