--- src/test/java/com/vaadin/flow/component/combobox/test/ComboboxSerializableTest.java
@@ -1,6 +1,0 @@
-package com.vaadin.flow.component.combobox.test;
-
-import com.vaadin.flow.testutil.ClassesSerializableTest;
-
-public class ComboboxSerializableTest extends ClassesSerializableTest {
-}
