+++ vaadin-combo-box-flow-integration-tests/src/main/java/com/vaadin/flow/component/combobox/bean/TestItem.java
