+++ vaadin-combo-box-flow-integration-tests/src/main/java/com/vaadin/flow/component/combobox/test/StringItemsWithTextRendererPage.java
