+++ vaadin-combo-box-flow-integration-tests/src/main/webapp/frontend/src/combo-box-in-a-template2.html
