+++ vaadin-combo-box-flow-integration-tests/src/test/java/com/vaadin/flow/component/combobox/test/template/ComboBoxInTemplateIT.java
