+++ vaadin-combo-box-flow/src/test/java/com/vaadin/flow/component/combobox/ComboboxSerializableTest.java
@@ -1,0 +1,6 @@
+package com.vaadin.flow.component.combobox;
+
+import com.vaadin.flow.testutil.ClassesSerializableTest;
+
+public class ComboboxSerializableTest extends ClassesSerializableTest {
+}
